feat: request and response mapping for inner/connector fetch notes
This change inspects subgraph requests from the inner query plan and constructs a vec of requests for the SubgraphConnector to call.

It handles:
* root fields (including aliased fields)
* entity resolvers
* fields on entities (by basically providing a stub entity resolver)

For each request, we collect the info necessary for mapping the response (JSON selection transforms and key path). The response handling function applies this information to construct a single response from multiple JSON objects.