feat: support cross-service joins
This is a hacky but effective way to make the "inner" supergraph accept `_entities` requests. For each entity with a connector, we generate a special root field we can use instead of `_entities`. Before query planning, we replace "_entities" with the root field name and shove it in the context. When we construct the inner responses, we use the root field name in the response data. And when we construct the outer response, we replace the root field naname with "_entities" again.