fix: "upgrade" join__type
Instead of adding redundant directives, this change "upgrades" the join__type directive, ensuring that there's only one per unique `graph:`. This is prep work for adding isInterfaceObject.

It also removes support for resolvable: false — if an entity isn't resolvable, it can't be a connector, so we don't care about it in the endpoint subgraph