test: tests for @requires and internal dependencies
Using @requires worked without any extra effort. We're really relying on the query planner to ensure that the entity representations include the required fields. It doesn't matter in the connector supergraph knows about these fields because we'll accept the `representations` variable as-is and make those values available to URLs and request bodies.

I also added a test where a field depends on other fields but they're not `@key` fields. This also worked without any additional effort.

All of the behavior really needs validation at composition time! I can imagine that we'll enforce additional restrictions at that point but we'll get to that later.

This is also the first version of the test supergraph built with new typescript composition, so yay.