feat: response formatting for aliases and abstract types (#314)
* feat: response formatting for aliases and abstract types

This change adds a version of the GraphQL execution algorithm to format connector JSON responses to match the incoming GraphQL operation. This occurs after applying the JSONSelection transforms, and before the router's existing response formatting.

It has special logic to handle aliases, ensuring that fields from the JSON are mapped appropriately to one or more response keys.

It also has special logic to determining a `__typename` for abstract types. For now, the logic is a heuristic that counts matching fields between the response object and the possible types. Eventually we'll add support for `keyTypeMap` and user-defined discriminators.

This change includes some operation manipulation code for situations where the "magic finder" field is used during query planning. Fortunately, that code is pretty well encapsulated in subgraph_connector and request_response.

---------

Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>