Lb/rest connectors/magic finder fix (#335)
* fix: rewrite field back to _entities before leaving inner supergraph

* fix: new way of handling magic finder fields

Prior to this fix, we made the simplifying assumption that we could generate one "finder"-style field for each entity type in the inner supergraph. However, some connectors on entity types had dependencies on other connectors, so they're not directly queryable from the outer query plan. The one-finder-field-per-type approach resulted in calls to connectors without their dependencies, which failed at runtime.

This new approach generates finder fields based on the Entity type name and the dependencies. So @sourceType on Product @key(fields: "id") results in Query._Product_id. Before we generate an inner query plan, we use the inputs to the fetch node (called requires in the node struct) to pick the most relevant finder field. This has a few benefits:

- If a connector has more dependencies (from a @requires, for example), it will have a more specific finder field. The query planner will not plan to that connector directly. Instead, it will generate a Sequence so the dependencies are fetched first.
- If two connectors share the same dependencies, they'll share the finder field. This lets the query planner generate a Parallel around the connector Fetch nodes.

The heuristics for picking the finder field name are probably incomplete, and it relies on the directives being accurate and coherent. It's possible for it to pick the wrong field name, which will result in a query planning error. I don't believe we can easily solve this without making a connector-aware Federation.