Remove the in-between types for Field, InlineFragment, and FragmentSpread
This means: no key caching, and you can modify things that modify the
key, if you have a mutable reference to a FieldSelection for example.

Note I'm pretty sure this is actually completely okay. We do not hand
out mutable references to selections in a selection map. The only
downside is not having a cached key. I think we can address that differently.